+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/generator/InvoiceLineGenerator.java
@@ -24,6 +24,7 @@ import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.Tax;
 import com.axelor.apps.account.db.TaxEquiv;
 import com.axelor.apps.account.db.TaxLine;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.exception.IExceptionMessage;
 import com.axelor.apps.account.service.AccountManagementAccountService;
 import com.axelor.apps.account.service.app.AppAccountService;
@@ -83,7 +84,9 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
   protected int discountTypeSelect;
   protected BigDecimal exTaxTotal;
   protected BigDecimal inTaxTotal;
-  protected boolean isTitleLine;
+  protected Integer typeSelect = 0;
+  protected boolean isSubLine = false;
+  protected Integer packPriceSelect = 0;
 
   public static final int DEFAULT_SEQUENCE = 0;
 
@@ -106,7 +109,9 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
       BigDecimal qty,
       Unit unit,
       int sequence,
-      boolean isTaxInvoice) {
+      boolean isTaxInvoice,
+      boolean isSubLine,
+      Integer packPriceSelect) {
 
     this(invoice);
 
@@ -119,6 +124,8 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
     this.isTaxInvoice = isTaxInvoice;
     this.today = Beans.get(AppAccountService.class).getTodayDate();
     this.currencyService = new CurrencyService(this.today);
+    this.isSubLine = isSubLine;
+    this.packPriceSelect = packPriceSelect;
   }
 
   protected InvoiceLineGenerator(
@@ -136,9 +143,11 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
       int discountTypeSelect,
       BigDecimal exTaxTotal,
       BigDecimal inTaxTotal,
-      boolean isTaxInvoice) {
+      boolean isTaxInvoice,
+      boolean isSubLine,
+      Integer packPriceSelect) {
 
-    this(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice);
+    this(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice, isSubLine, packPriceSelect);
 
     this.price = price;
     this.priceDiscounted = priceDiscounted;
@@ -168,13 +177,14 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 
     InvoiceLine invoiceLine = new InvoiceLine();
     boolean isPurchase = InvoiceToolService.isPurchase(invoice);
+    boolean accountRequired = isAccountRequired();
 
     invoiceLine.setInvoice(invoice);
 
     invoiceLine.setProduct(product);
 
     invoiceLine.setProductName(productName);
-    if (product != null) {
+    if (product != null && accountRequired) {
       invoiceLine.setProductCode(product.getCode());
       AccountManagement accountManagement =
           accountManagementService.getAccountManagement(product, invoice.getCompany());
@@ -185,15 +195,19 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
     invoiceLine.setDescription(description);
     invoiceLine.setPrice(price);
 
+    invoiceLine.setIsSubLine(isSubLine);
     invoiceLine.setPriceDiscounted(priceDiscounted);
     invoiceLine.setQty(qty);
     invoiceLine.setUnit(unit);
 
-    if (taxLine == null) {
+    invoiceLine.setTypeSelect(typeSelect);
+    invoiceLine.setPackPriceSelect(packPriceSelect);
+
+    if (taxLine == null && accountRequired) {
       this.determineTaxLine();
     }
 
-    if (product != null) {
+    if (product != null && accountRequired) {
       Tax tax =
           Beans.get(AccountManagementService.class)
               .getProductTax(
@@ -228,9 +242,11 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
     invoiceLine.setDiscountTypeSelect(discountTypeSelect);
     invoiceLine.setDiscountAmount(discountAmount);
 
-    invoiceLine.setIsTitleLine(isTitleLine);
-
     return invoiceLine;
+  }
+
+  public boolean isAccountRequired() {
+    return true;
   }
 
   public void determineTaxLine() throws AxelorException {
@@ -252,7 +268,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 
   public void computeTotal() {
 
-    if (isTitleLine) {
+    if (typeSelect == InvoiceLineRepository.TYPE_TITLE) {
       return;
     }
 
@@ -272,7 +288,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 
   public void computeCompanyTotal(InvoiceLine invoiceLine) throws AxelorException {
 
-    if (isTitleLine) {
+    if (typeSelect == InvoiceLineRepository.TYPE_TITLE) {
       return;
     }