+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/expense/ExpenseServiceImpl.java
@@ -619,7 +619,9 @@ public class ExpenseServiceImpl implements ExpenseService {
               PriceListLineRepository.AMOUNT_TYPE_NONE,
               expenseLine.getUntaxedAmount(),
               expenseLine.getTotalAmount(),
-              false) {
+              false,
+              false,
+              0) {
 
             @Override
             public List<InvoiceLine> creates() throws AxelorException {
@@ -649,7 +651,9 @@ public class ExpenseServiceImpl implements ExpenseService {
               PriceListLineRepository.AMOUNT_TYPE_NONE,
               expenseLine.getUntaxedAmount(),
               expenseLine.getTotalAmount(),
-              false) {
+              false,
+              false,
+              0) {
 
             @Override
             public List<InvoiceLine> creates() throws AxelorException {