+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/configurator/ConfiguratorServiceImpl.java
@@ -221,7 +221,8 @@ public class ConfiguratorServiceImpl implements ConfiguratorService {
   protected void fillSaleOrderWithProduct(SaleOrderLine saleOrderLine) throws AxelorException {
     SaleOrderLineService saleOrderLineService = Beans.get(SaleOrderLineService.class);
     if (saleOrderLine.getProduct() != null) {
-      saleOrderLineService.computeProductInformation(saleOrderLine, saleOrderLine.getSaleOrder());
+      saleOrderLineService.computeProductInformation(
+          saleOrderLine, saleOrderLine.getSaleOrder(), null);
     }
   }