+++ axelor-sale/src/main/java/com/axelor/apps/sale/web/SaleOrderLineController.java
@@ -34,6 +34,7 @@ import com.axelor.rpc.Context;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.List;
 import java.util.Map;
 
@@ -53,14 +54,7 @@ public class SaleOrderLineController {
     SaleOrder saleOrder = saleOrderLineService.getSaleOrder(context);
 
     try {
-      Map<String, BigDecimal> map = saleOrderLineService.computeValues(saleOrder, saleOrderLine);
-
-      response.setValues(map);
-      response.setAttr(
-          "priceDiscounted",
-          "hidden",
-          map.getOrDefault("priceDiscounted", BigDecimal.ZERO).compareTo(saleOrderLine.getPrice())
-              == 0);
+      compute(response, saleOrder, saleOrderLine);
     } catch (Exception e) {
       TraceBackService.trace(response, e);
     }
@@ -91,6 +85,7 @@ public class SaleOrderLineController {
       Context context = request.getContext();
       SaleOrderLine saleOrderLine = context.asType(SaleOrderLine.class);
       SaleOrder saleOrder = saleOrderLineService.getSaleOrder(context);
+
       Product product = saleOrderLine.getProduct();
 
       if (saleOrder == null || product == null) {
@@ -98,8 +93,18 @@ public class SaleOrderLineController {
         return;
       }
 
+      Integer packPriceSelect = product.getPackPriceSelect();
+      if (saleOrderLine.getIsSubLine()) {
+        if (context.getParent().getContextClass().equals(SaleOrderLine.class)) {
+          packPriceSelect = context.getParent().asType(SaleOrderLine.class).getPackPriceSelect();
+        } else if (saleOrderLine.getParentLine() != null) {
+          packPriceSelect = saleOrderLine.getParentLine().getPackPriceSelect();
+        }
+      }
+
       try {
-        saleOrderLineService.computeProductInformation(saleOrderLine, saleOrder);
+        product = Beans.get(ProductRepository.class).find(product.getId());
+        saleOrderLineService.computeProductInformation(saleOrderLine, saleOrder, packPriceSelect);
         response.setValue("saleSupplySelect", product.getSaleSupplySelect());
         response.setValues(saleOrderLine);
       } catch (Exception e) {
@@ -115,6 +120,9 @@ public class SaleOrderLineController {
     Beans.get(SaleOrderLineService.class).resetProductInformation(line);
     response.setValue("saleSupplySelect", null);
     response.setValues(line);
+    response.setValue("typeSelect", SaleOrderLineRepository.TYPE_NORMAL);
+    response.setValue("packPriceSelect", null);
+    response.setValue("subLineList", null);
   }
 
   public void getTaxEquiv(ActionRequest request, ActionResponse response) {
@@ -210,36 +218,12 @@ public class SaleOrderLineController {
   public void emptyLine(ActionRequest request, ActionResponse response) {
     SaleOrderLine saleOrderLine = request.getContext().asType(SaleOrderLine.class);
     if (saleOrderLine.getTypeSelect() != SaleOrderLineRepository.TYPE_NORMAL) {
-      SaleOrderLine newSaleOrderLine = new SaleOrderLine();
-      newSaleOrderLine.setQty(BigDecimal.ZERO);
-      newSaleOrderLine.setId(saleOrderLine.getId());
-      newSaleOrderLine.setVersion(saleOrderLine.getVersion());
-      newSaleOrderLine.setTypeSelect(saleOrderLine.getTypeSelect());
-      response.setValues(Mapper.toMap(newSaleOrderLine));
-    }
-  }
-
-  public void createPackLines(ActionRequest request, ActionResponse response)
-      throws AxelorException {
-
-    SaleOrderLine soLine = request.getContext().asType(SaleOrderLine.class);
-
-    Product product = soLine.getProduct();
-
-    if (product != null) {
-
-      product = Beans.get(ProductRepository.class).find(product.getId());
-
-      if (product.getIsPack()) {
-        SaleOrder saleOrder = saleOrderLineService.getSaleOrder(request.getContext());
-        List<SaleOrderLine> subLines = saleOrderLineService.createPackLines(product, saleOrder);
-
-        if (!subLines.isEmpty()) {
-          response.setValue("subLineList", subLines);
-        }
-        response.setValue("typeSelect", SaleOrderLineRepository.TYPE_PACK);
-        response.setValue("qty", 0);
-      }
+      Map<String, Object> newSaleOrderLine = Mapper.toMap(new SaleOrderLine());
+      newSaleOrderLine.put("qty", BigDecimal.ZERO);
+      newSaleOrderLine.put("id", saleOrderLine.getId());
+      newSaleOrderLine.put("version", saleOrderLine.getVersion());
+      newSaleOrderLine.put("typeSelect", saleOrderLine.getTypeSelect());
+      response.setValues(newSaleOrderLine);
     }
   }
 
@@ -248,5 +232,132 @@ public class SaleOrderLineController {
     Context context = request.getContext();
     SaleOrderLine saleOrderLine = context.asType(SaleOrderLine.class);
     saleOrderLineService.checkMultipleQty(saleOrderLine, response);
+  }
+
+  public void updateSubLineQty(ActionRequest request, ActionResponse response)
+      throws AxelorException {
+
+    SaleOrderLine newkitLine = request.getContext().asType(SaleOrderLine.class);
+    BigDecimal qty = BigDecimal.ZERO;
+    BigDecimal oldKitQty = BigDecimal.ZERO;
+    BigDecimal newKitQty = BigDecimal.ZERO;
+    BigDecimal exTaxTotal = BigDecimal.ZERO;
+    BigDecimal inTaxTotal = BigDecimal.ZERO;
+    BigDecimal priceDiscounted = BigDecimal.ZERO;
+    BigDecimal taxRate = BigDecimal.ZERO;
+    ;
+    BigDecimal companyExTaxTotal = BigDecimal.ZERO;
+    BigDecimal companyInTaxTotal = BigDecimal.ZERO;
+
+    Context context = request.getContext();
+    SaleOrder saleOrder = saleOrderLineService.getSaleOrder(context);
+
+    if (newkitLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK) {
+
+      if (newkitLine.getOldQty().compareTo(BigDecimal.ZERO) == 0) {
+        oldKitQty = BigDecimal.ONE;
+      } else {
+        oldKitQty = newkitLine.getOldQty();
+      }
+
+      if (newkitLine.getQty().compareTo(BigDecimal.ZERO) != 0) {
+        newKitQty = newkitLine.getQty();
+      }
+
+      List<SaleOrderLine> orderLines = newkitLine.getSubLineList();
+
+      if (orderLines != null) {
+        if (newKitQty.compareTo(BigDecimal.ZERO) != 0) {
+          for (SaleOrderLine line : orderLines) {
+            qty = (line.getQty().divide(oldKitQty, 2, RoundingMode.HALF_EVEN)).multiply(newKitQty);
+            priceDiscounted = saleOrderLineService.computeDiscount(line);
+
+            if (line.getTaxLine() != null) {
+              taxRate = line.getTaxLine().getValue();
+            }
+
+            if (!saleOrder.getInAti()) {
+              exTaxTotal = saleOrderLineService.computeAmount(qty, priceDiscounted);
+              inTaxTotal = exTaxTotal.add(exTaxTotal.multiply(taxRate));
+              companyExTaxTotal =
+                  saleOrderLineService.getAmountInCompanyCurrency(exTaxTotal, saleOrder);
+              companyInTaxTotal = companyExTaxTotal.add(companyExTaxTotal.multiply(taxRate));
+            } else {
+              inTaxTotal = saleOrderLineService.computeAmount(qty, priceDiscounted);
+              exTaxTotal =
+                  inTaxTotal.divide(taxRate.add(BigDecimal.ONE), 2, BigDecimal.ROUND_HALF_UP);
+              companyInTaxTotal =
+                  saleOrderLineService.getAmountInCompanyCurrency(inTaxTotal, line.getSaleOrder());
+              companyExTaxTotal =
+                  companyInTaxTotal.divide(
+                      taxRate.add(BigDecimal.ONE), 2, BigDecimal.ROUND_HALF_UP);
+            }
+
+            line.setQty(qty.setScale(2, RoundingMode.HALF_EVEN));
+            line.setPriceDiscounted(priceDiscounted);
+            line.setExTaxTotal(exTaxTotal);
+            line.setInTaxTotal(inTaxTotal);
+            line.setCompanyExTaxTotal(companyExTaxTotal);
+            line.setCompanyInTaxTotal(companyInTaxTotal);
+          }
+        } else {
+          for (SaleOrderLine line : orderLines) {
+            line.setQty(qty.setScale(2, RoundingMode.HALF_EVEN));
+          }
+        }
+
+        response.setValue("oldQty", newKitQty);
+        response.setValue("subLineList", orderLines);
+      }
+    }
+  }
+
+  public void resetPackLine(ActionRequest request, ActionResponse response) throws AxelorException {
+
+    Context context = request.getContext();
+    SaleOrder saleOrder = saleOrderLineService.getSaleOrder(context);
+    SaleOrderLine packLine = context.asType(SaleOrderLine.class);
+    try {
+      saleOrderLineService.fillPrice(packLine, saleOrder, packLine.getPackPriceSelect());
+      compute(response, saleOrder, packLine);
+    } catch (Exception e) {
+      e.printStackTrace();
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  public void resetPackSubLine(ActionRequest request, ActionResponse response)
+      throws AxelorException {
+
+    Context context = request.getContext();
+    SaleOrder saleOrder = saleOrderLineService.getSaleOrder(context);
+    SaleOrderLine packLine = context.asType(SaleOrderLine.class);
+    List<SaleOrderLine> subLines = packLine.getSubLineList();
+
+    try {
+      if (subLines != null) {
+        for (SaleOrderLine line : subLines) {
+          saleOrderLineService.fillPrice(line, saleOrder, packLine.getPackPriceSelect());
+          saleOrderLineService.computeValues(saleOrder, line);
+        }
+        response.setValue("subLineList", subLines);
+      }
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  private void compute(ActionResponse response, SaleOrder saleOrder, SaleOrderLine orderLine)
+      throws AxelorException {
+
+    Map<String, BigDecimal> map = saleOrderLineService.computeValues(saleOrder, orderLine);
+    map.put("price", orderLine.getPrice());
+    map.put("companyCostPrice", orderLine.getCompanyCostPrice());
+    map.put("discountAmount", orderLine.getDiscountAmount());
+    response.setValues(map);
+    response.setAttr(
+        "priceDiscounted",
+        "hidden",
+        map.getOrDefault("priceDiscounted", BigDecimal.ZERO).compareTo(orderLine.getPrice()) == 0);
   }
 }