+++ axelor-stock/src/main/java/com/axelor/apps/stock/web/StockMoveLineController.java
@@ -19,6 +19,7 @@ package com.axelor.apps.stock.web;
 
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.StockMoveLine;
+import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
 import com.axelor.apps.stock.service.StockMoveLineService;
 import com.axelor.db.mapper.Mapper;
 import com.axelor.exception.AxelorException;
@@ -29,6 +30,8 @@ import com.axelor.rpc.ActionResponse;
 import com.axelor.rpc.Context;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import java.math.BigDecimal;
+import java.util.Map;
 
 @Singleton
 public class StockMoveLineController {
@@ -40,8 +43,11 @@ public class StockMoveLineController {
     StockMove stockMove = stockMoveLine.getStockMove();
     if (stockMove == null) {
       Context parentContext = request.getContext().getParent();
+      Context superParentContext = parentContext.getParent();
       if (parentContext.getContextClass().equals(StockMove.class)) {
         stockMove = parentContext.asType(StockMove.class);
+      } else if (superParentContext.getContextClass().equals(StockMove.class)) {
+        stockMove = superParentContext.asType(StockMove.class);
       } else {
         return;
       }
@@ -69,6 +75,18 @@ public class StockMoveLineController {
       stockMoveLine = new StockMoveLine();
       response.setValues(Mapper.toMap(stockMoveLine));
       TraceBackService.trace(response, e, ResponseMessageType.INFORMATION);
+    }
+  }
+
+  public void emptyLine(ActionRequest request, ActionResponse response) {
+    StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+    if (stockMoveLine.getLineTypeSelect() != StockMoveLineRepository.TYPE_NORMAL) {
+      Map<String, Object> newStockMoveLine = Mapper.toMap(new StockMoveLine());
+      newStockMoveLine.put("qty", BigDecimal.ZERO);
+      newStockMoveLine.put("id", stockMoveLine.getId());
+      newStockMoveLine.put("version", stockMoveLine.getVersion());
+      newStockMoveLine.put("lineTypeSelect", stockMoveLine.getLineTypeSelect());
+      response.setValues(newStockMoveLine);
     }
   }
 }