+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderInvoiceService.java
@@ -300,4 +300,5 @@ public interface SaleOrderInvoiceService {
    * @return the domain for the operation select field in the invoicing wizard form
    */
   Map<String, Integer> getInvoicingWizardOperationDomain(SaleOrder saleOrder);
+
 }