+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/invoice/generator/InvoiceLineGeneratorSupplyChain.java
@@ -24,6 +24,7 @@ import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.TaxLine;
 import com.axelor.apps.account.db.repo.AnalyticMoveLineMngtRepository;
 import com.axelor.apps.account.db.repo.AnalyticMoveLineRepository;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.service.invoice.InvoiceLineService;
 import com.axelor.apps.account.service.invoice.InvoiceToolService;
 import com.axelor.apps.account.service.invoice.generator.InvoiceLineGenerator;
@@ -32,7 +33,7 @@ import com.axelor.apps.base.db.Unit;
 import com.axelor.apps.base.service.UnitConversionService;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
 import com.axelor.apps.sale.db.SaleOrderLine;
-import com.axelor.apps.sale.db.repo.SaleOrderLineRepository;
+import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
@@ -68,7 +69,9 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
       boolean isTaxInvoice,
       SaleOrderLine saleOrderLine,
       PurchaseOrderLine purchaseOrderLine,
-      StockMoveLine stockMoveLine) {
+      StockMoveLine stockMoveLine,
+      boolean isSubLine,
+      Integer packPriceSelect) {
     super(
         invoice,
         product,
@@ -84,7 +87,9 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
         discountTypeSelect,
         exTaxTotal,
         inTaxTotal,
-        isTaxInvoice);
+        isTaxInvoice,
+        isSubLine,
+        packPriceSelect);
     this.saleOrderLine = saleOrderLine;
     this.purchaseOrderLine = purchaseOrderLine;
     this.stockMoveLine = stockMoveLine;
@@ -101,10 +106,12 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
       boolean isTaxInvoice,
       SaleOrderLine saleOrderLine,
       PurchaseOrderLine purchaseOrderLine,
-      StockMoveLine stockMoveLine)
+      StockMoveLine stockMoveLine,
+      boolean isSubLine,
+      Integer packPriceSelect)
       throws AxelorException {
 
-    super(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice);
+    super(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice, isSubLine, packPriceSelect);
 
     this.saleOrderLine = saleOrderLine;
     this.purchaseOrderLine = purchaseOrderLine;
@@ -116,9 +123,11 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
       this.priceDiscounted = saleOrderLine.getPriceDiscounted();
       this.taxLine = saleOrderLine.getTaxLine();
       this.discountTypeSelect = saleOrderLine.getDiscountTypeSelect();
-      this.isTitleLine = saleOrderLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK;
+      this.typeSelect = saleOrderLine.getTypeSelect();
     } else if (purchaseOrderLine != null) {
-      this.isTitleLine = purchaseOrderLine.getIsTitleLine();
+      if (purchaseOrderLine.getIsTitleLine()) {
+        this.typeSelect = InvoiceLineRepository.TYPE_TITLE;
+      }
       this.purchaseOrderLine = purchaseOrderLine;
       this.discountAmount = purchaseOrderLine.getDiscountAmount();
       this.price = purchaseOrderLine.getPrice();
@@ -250,5 +259,25 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
     } else {
       return product.getPurchasesUnit();
     }
+  }
+
+  @Override
+  public boolean isAccountRequired() {
+
+    if (Beans.get(AppSaleService.class).getAppSale().getProductPackMgt()) {
+
+      if (isSubLine
+          && saleOrderLine.getParentLine() != null
+          && saleOrderLine.getParentLine().getPackPriceSelect()
+              == InvoiceLineRepository.PACK_PRICE_ONLY) {
+        return false;
+      }
+      if (typeSelect == InvoiceLineRepository.TYPE_PACK
+          && saleOrderLine.getPackPriceSelect() == InvoiceLineRepository.SUBLINE_PRICE_ONLY) {
+        return false;
+      }
+    }
+
+    return true;
   }
 }